requirements.txt
ansys-api-dbu==0.3.28
ansys-api-fluent==0.3.36
ansys-api-geometry==0.4.83
ansys-api-platform-instancemanagement==1.1.3
ansys-api-tools-filetransfer==0.1.1
ansys-fluent-core==0.35.0
ansys-geometry-core==0.12.0
ansys-platform-instancemanagement==1.1.2
ansys-tools-filetransfer==0.1.1
ansys-tools-path==0.7.3
ansys-units==0.8.1
beartype==0.22.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
defusedxml==0.7.1
Deprecated==1.2.18
docker==7.1.0
et_xmlfile==2.0.0
flexcache==0.3
flexparser==0.4
fonttools==4.60.1
geomdl==5.4.0
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-health-checking==1.62.3
grpcio-status==1.62.3
idna==3.11
importlib_metadata==8.7.0
joblib==1.5.2
kiwisolver==1.4.9
matplotlib==3.10.7
nltk==3.9.2
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
Pint==0.25
platformdirs==4.5.0
protobuf==4.25.8
pyansys-tools-report==0.8.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
scipy==1.16.2
scooby==0.10.2
semver==3.0.4
setuptools==80.9.0
six==1.17.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wrapt==1.17.3
xlsxwriter==3.2.9
zipp==3.23.0
